:lock: secure initialization

+++ api/initialization.go
@@ -3,11 +3,18 @@ package api
 import (
 	"hackz-allo/db"
 	"net/http"
+	"os"
 
 	"github.com/labstack/echo/v4"
 )
 
 func Initialization(c echo.Context) error {
+
+	pw := c.QueryParam("pw")
+
+	if pw != os.Getenv("INIT_PASSWORD") {
+		return c.String(http.StatusOK, "Failed")
+	}
 
 	db.Psql.Exec("DROP TABLE IF EXISTS users")
 	db.Psql.Exec("DROP TABLE IF EXISTS posts")
@@ -17,5 +24,5 @@ func Initialization(c echo.Context) error {
 	db.Psql.AutoMigrate(db.Post{})
 	db.Psql.AutoMigrate(db.Friend{})
 
-	return c.String(http.StatusOK, "Initializaton")
+	return c.String(http.StatusOK, "Successed")
 }
